agentic_ai_project/requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-groq
langchain-mistralai
langchain-openai
langchain-text-splitters
faiss-cpu
PyPDF2
google-generativeai
openai
tavily-python
langchain-tavily
gTTS
google-api-python-client
google-auth
requests
huggingface-hub